+++ tests/model/test_ocrd_page.py
@@ -8,6 +8,7 @@ from ocrd_models.ocrd_page import (
     TextRegionType,
     TextLineType,
     OrderedGroupIndexedType,
+    UnorderedGroupIndexedType,
     RegionRefIndexedType,
     WordType,
     GlyphType,
@@ -151,19 +152,19 @@ class TestOcrdPage(TestCase):
             og = pcgts.get_Page().get_ReadingOrder().get_OrderedGroup()
             xml_before = to_xml(og)
             children = og.get_AllIndexed()
-            self.assertEqual(len(children), 20)
-            self.assertEqual([c.index for c in children], list(range(0, 20)))
+            self.assertEqual(len(children), 22)
+            self.assertEqual([c.index for c in children], list(range(0, 22)))
             # mix up the indexes
             children[0].index = 11
             children[11].index = 3
             children[3].index = 0
-            self.assertEqual([c.index for c in children], [11, 1, 2, 0, 4, 5, 6, 7, 8, 9, 10, 3, 12, 13, 14, 15, 16, 17, 18, 19])
-            self.assertEqual([c.index for c in og.get_AllIndexed()], list(range(0, 20)))
+            self.assertEqual([c.index for c in children], [11, 1, 2, 0, 4, 5, 6, 7, 8, 9, 10, 3, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21])
+            self.assertEqual([c.index for c in og.get_AllIndexed()], list(range(0, 22)))
             self.assertEqual(og.get_AllIndexed()[1].__class__, OrderedGroupIndexedType)
             # serialize and make sure the correct order was serialized
             new_pcgts = parseString(to_xml(pcgts).encode('utf8'), silence=True)
             new_og = new_pcgts.get_Page().get_ReadingOrder().get_OrderedGroup()
-            self.assertEqual([c.index for c in new_og.get_AllIndexed()], list(range(0, 20)))
+            self.assertEqual([c.index for c in new_og.get_AllIndexed()], list(range(0, 22)))
             # xml_after = to_xml(new_og)
             # self.assertEqual(xml_after, xml_before)
 
@@ -176,6 +177,7 @@ class TestOcrdPage(TestCase):
             og = pcgts.get_Page().get_ReadingOrder().get_OrderedGroup()
             children = og.get_AllIndexed()
             self.assertTrue(isinstance(children[1], OrderedGroupIndexedType))
+            self.assertTrue(isinstance(children[21], UnorderedGroupIndexedType))
             # empty all the elements in the first orederdGroupIndexed
             children[1].set_RegionRefIndexed([])
             # serialize apnd parse to see empty group converted
@@ -183,6 +185,93 @@ class TestOcrdPage(TestCase):
             og = pcgts.get_Page().get_ReadingOrder().get_OrderedGroup()
             children = og.get_AllIndexed()
             self.assertTrue(isinstance(children[1], RegionRefIndexedType))
+            self.assertTrue(isinstance(children[21], RegionRefIndexedType))
+
+    def test_all_regions_without_reading_order(self):
+        """
+        https://github.com/OCR-D/core/pull/479
+        https://github.com/OCR-D/core/issues/240#issuecomment-493135797
+        """
+        with open('tests/model/TEMP1_Gutachten2-2.xml', 'r') as f:
+            pcgts = parseString(f.read().encode('utf8'), silence=True)
+            pg = pcgts.get_Page()
+            self.assertEqual(len(pg.get_AllRegions()), 65)
+            self.assertEqual(len(pg.get_AllRegions(depth=0)), 65)
+            self.assertEqual(len(pg.get_AllRegions(depth=1)), 45)
+            self.assertEqual(len(pg.get_AllRegions(depth=2)), 65)
+            self.assertEqual(len(pg.get_AllRegions(depth=3)), 65)
+            self.assertEqual(len(pg.get_AllRegions(classes=['Separator'])), 25)
+            self.assertEqual(len(pg.get_AllRegions(classes=['Table'])), 3)
+            self.assertEqual(len(pg.get_AllRegions(classes=['Text'])), 37)
+            self.assertEqual(len(pg.get_AllRegions(classes=['Text'], depth=1)), 17)
+            self.assertEqual(len(pg.get_AllRegions(classes=['Text'], depth=2)), 37)
+
+    def test_all_regions_with_reading_order(self):
+        """
+        https://github.com/OCR-D/core/pull/479
+        https://github.com/OCR-D/core/issues/240#issuecomment-493135797
+        """
+        with open('tests/model/TEMP1_Gutachten2-2.xml', 'r') as f:
+            pg = parseString(f.read().encode('utf8'), silence=True).get_Page()
+            with self.assertRaisesRegex(Exception, "Argument 'order' must be either 'document', 'reading-order' or 'reading-order-only', not 'random'"):
+                pg.get_AllRegions(order='random')
+            with self.assertRaisesRegex(Exception, "Argument 'depth' must be an integer greater-or-equal 0, not '-1'"):
+                pg.get_AllRegions(depth=-1)
+            self.assertEqual(len(pg.get_AllRegions(order='reading-order-only')), 40)
+            self.assertEqual(len(pg.get_AllRegions(order='reading-order-only', depth=1)), 20)
+            self.assertEqual(len(pg.get_AllRegions(order='reading-order-only', depth=2)), 40)
+            self.assertEqual(len(pg.get_AllRegions(order='reading-order', depth=0)), 65)
+            self.assertEqual(len(pg.get_AllRegions(order='reading-order', depth=1)), 45)
+            self.assertEqual(len(pg.get_AllRegions(order='reading-order', depth=2)), 65)
+            self.assertEqual(len(pg.get_AllRegions(classes=['Table'], order='reading-order')), 3)
+            self.assertEqual(len(pg.get_AllRegions(classes=['Text'], order='reading-order')), 37)
+            self.assertEqual(len(pg.get_AllRegions(classes=['Text'], order='reading-order', depth=1)), 17)
+
+    def test_get_UnorderdGroupChildren(self):
+        with open('tests/model/TEMP1_Gutachten2-2.xml', 'r') as f:
+            pcgts = parseString(f.read().encode('utf8'), silence=True)
+            ug = pcgts.get_Page().get_ReadingOrder().get_OrderedGroup().get_UnorderedGroupIndexed()[0]
+            self.assertEqual(len(ug.get_UnorderedGroupChildren()), 1)
+
+    def test_get_AllIndexed_classes(self):
+        with open('tests/model/TEMP1_Gutachten2-2.xml', 'r') as f:
+            og = parseString(f.read().encode('utf8'), silence=True).get_Page().get_ReadingOrder().get_OrderedGroup()
+            self.assertEqual(len(og.get_AllIndexed(classes=['RegionRef'])), 17)
+            self.assertEqual(len(og.get_AllIndexed(classes=['OrderedGroup'])), 3)
+            self.assertEqual(len(og.get_AllIndexed(classes=['UnorderedGroup'])), 2)
+
+    def test_get_AllIndexed_index_sort(self):
+        with open('tests/model/TEMP1_Gutachten2-2.xml', 'r') as f:
+            og = parseString(f.read().encode('utf8'), silence=True).get_Page().get_ReadingOrder().get_OrderedGroup()
+            unogs = og.get_UnorderedGroupIndexed()
+            self.assertEqual([x.index for x in unogs], [20, 21])
+            unogs[0].index = 21
+            unogs[1].index = 20
+            self.assertEqual([x.index for x in og.get_AllIndexed(classes=['UnorderedGroup'], index_sort=True)], [20, 21])
+            self.assertEqual([x.index for x in og.get_AllIndexed(classes=['UnorderedGroup'], index_sort=False)], [21, 20])
+            og.sort_AllIndexed()
+            self.assertEqual([x.index for x in og.get_AllIndexed(classes=['UnorderedGroup'], index_sort=False)], [20, 21])
+
+    def test_extend_AllIndexed_no_validation(self):
+        with open('tests/model/TEMP1_Gutachten2-2.xml', 'r') as f:
+            og = parseString(f.read().encode('utf8'), silence=True).get_Page().get_ReadingOrder().get_OrderedGroup()
+            og.extend_AllIndexed([
+                RegionRefIndexedType(index=3, id='r3'),
+                RegionRefIndexedType(index=2, id='r2'),
+                RegionRefIndexedType(index=1, id='r1'),
+            ])
+            rrs = og.get_RegionRefIndexed()
+            self.assertEqual([x.index for x in rrs][-3:], [22, 23, 24])
+
+    def test_extend_AllIndexed_validate_continuity(self):
+        with open('tests/model/TEMP1_Gutachten2-2.xml', 'r') as f:
+            og = parseString(f.read().encode('utf8'), silence=True).get_Page().get_ReadingOrder().get_OrderedGroup()
+            with self.assertRaisesRegex(Exception, "@index already used: 1"):
+                og.extend_AllIndexed([
+                    RegionRefIndexedType(index=3, id='r3'),
+                    RegionRefIndexedType(index=2, id='r2'),
+                    RegionRefIndexedType(index=1, id='r1'),
+                ], validate_continuity=True)
 
 if __name__ == '__main__':
     main()